requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlmodel==0.0.22
SQLAlchemy==2.0.36
pydantic-settings==2.7.0
python-jose[cryptography]==3.3.0
bcrypt==5.0.0
python-multipart==0.0.20
httpx==0.28.1
psycopg2-binary
alembic
cachetools
openai
prometheus-client
pytest
pytest-asyncio
python-json-logger
slowapi
tenacity
